完善店铺管理；完善店铺商品（还需要优化功能）；

+++ api/src/main/java/cn/qihangerp/api/domain/SysShopPullLasttime.java
@@ -19,7 +19,7 @@ public class SysShopPullLasttime implements Serializable {
     /**
      * 店铺id
      */
-    private Integer shopId;
+    private Long shopId;
 
     /**
      * 类型（ORDER:订单，REFUND:退款）
